EKTR Navisworks Catalog
Initial commit - the file was prepared for educational purposes based on the available ÉKTR (AZ ÉPÍTÉSI BERUHÁZÁSOK KÖLTSÉGTERVEZÉSI RENDSZERE) classification developed for the Hungarian AECO industry. It contains all available classes as Groups, not using Items at the moment - this is to be revised in the future. Baseline units are scarcely set, colours and transparency are not developed. Use the Excel Macro to generate the XML file, do not edit it directly.